ObjSegAD/requirements.txt
# Core dependencies
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.19.0
pillow>=8.0.0
opencv-python>=4.5.0
scikit-learn>=0.24.0
matplotlib>=3.3.0

# Data processing
pandas>=1.2.0
scipy>=1.6.0

# Image processing
albumentations>=1.0.0
imageio>=2.9.0

# Utilities
tqdm>=4.60.0
pyyaml>=5.4.0
tensorboard>=2.5.0

# Optional: for advanced features
# jupyter>=1.0.0
# seaborn>=0.11.0
# plotly>=5.0.0